third_party/blink/web_tests/fast/forms/search/search-cancel-button-events-expected.txt

Test for event dipatching by search cancel button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial state
PASS changeEventCounter is 0
PASS inputEventCounter is 0
Click the cancel button
PASS search.value is ""
PASS changeEventCounter is 0
PASS inputEventCounter is 1
Focus on another field
PASS changeEventCounter is 1
PASS inputEventCounter is 1
PASS successfullyParsed is true

TEST COMPLETE
